+++ src/collective/feedback/testing.py
@@ -1,16 +1,17 @@
-from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
-from plone.app.testing import applyProfile
-from plone.app.testing import FunctionalTesting
-from plone.app.testing import IntegrationTesting
-from plone.app.testing import PloneSandboxLayer
-from plone.testing.zope import WSGI_SERVER_FIXTURE
-
-import collective.feedback
 import collective.honeypot
 import collective.honeypot.config
 import plone.restapi
 import souper.plone
+from plone.app.contenttypes.testing import PLONE_APP_CONTENTTYPES_FIXTURE
+from plone.app.testing import (
+    FunctionalTesting,
+    IntegrationTesting,
+    PloneSandboxLayer,
+    applyProfile,
+)
+from plone.testing.zope import WSGI_SERVER_FIXTURE
 
+import collective.feedback
 
 collective.honeypot.config.EXTRA_PROTECTED_ACTIONS = set(["feedback-add"])
 collective.honeypot.config.HONEYPOT_FIELD = "honey"